// frontend/src/types/types.ts
export interface ISong {
  _id?: string; // make _id optional
  title: string;
  artists: string[];
  genres: string[];
  album?: string; // make album optional
}

export interface Song {
  _id: string;
  title: string;
  artists: { _id: string; name: string }[];
  genres: { _id: string; name: string }[];
  album: { _id: string; name: string } | null;
  fileUrl: string;
  createdAt: string;
  updatedAt: string;
}

export interface IArtist {
  _id?: string; // Make _id optional if it is not provided during creation
  name: string;
}

export interface Artist {
  _id: string;
  name: string;
}

export interface IGenre {
  _id?: string; // Make _id optional if it is not required at creation
  name: string;
}

export interface Genre {
  _id: string;
  name: string;
}

export interface IAlbum {
  _id?: string; // Make _id optional if it's generated by the backend
  name: string;
  releaseDate?: string; // Optional field for release date
}

export interface Album {
  _id: string;
  name: string;
  songs: string[]; // Array of song IDs
  artists: string[]; // Array of artist IDs
  createdAt: string;
  updatedAt: string;
}